+++ src/test_encoder.cpp
@@ -35,7 +35,7 @@ constexpr unsigned width = 640;
 int main(int argc, char **argv) {
   std::cout << std::endl << "Welcome to the H264 encoder test" << std::endl << std::endl;
   Encoder encoder(100000, width, height, 25);
-  for (size_t i = 0; i < 100; i++) {
+  for (unsigned i = 0; i < 100; i++) {
     Image raw_image = generate_strip_image(i, i + 10, width, height);
     auto data = encoder.encode(raw_image.buffer.data());
     if (data.size()) {
